rename to _HtmlText
